chore: moved thermoextrap.legacy -> tests/legacy
